copr: refactor code
