+++ components/tidb_query/src/expr/builtin_math.rs
@@ -10,7 +10,7 @@ use tikv_util::file::calc_crc32_bytes;
 use super::{Error, EvalContext, Result, ScalarFunc};
 use crate::codec::mysql::{Decimal, RoundMode, DEFAULT_FSP};
 use crate::codec::Datum;
-use crate::util::MySQLRng;
+use crate::expr_util::rand::MySQLRng;
 
 impl ScalarFunc {
     #[inline]
