+++ components/tidb_query/src/expr_util/mod.rs
@@ -5,3 +5,4 @@
 pub mod conv;
 pub mod like;
 pub mod miscellaneous;
+pub mod rand;
